Refactor code formatting and linting
